changed feature_estimates from integer to float

+++ db/migrate/20190711222004_add_column_to_feature_estimates.rb
@@ -1,10 +1,10 @@
 class AddColumnToFeatureEstimates < ActiveRecord::Migration[5.2]
   def change
-    add_column :feature_estimates, :design, :integer
-    add_column :feature_estimates, :qaTesting, :integer
-    add_column :feature_estimates, :deployment, :integer
-    add_column :feature_estimates, :postDeploymentDev, :integer
-    add_column :feature_estimates, :projectManagement, :integer
-    add_column :feature_estimates, :generalBuffer, :integer
+    add_column :feature_estimates, :design, :float
+    add_column :feature_estimates, :qaTesting, :float
+    add_column :feature_estimates, :deployment, :float
+    add_column :feature_estimates, :postDeploymentDev, :float
+    add_column :feature_estimates, :projectManagement, :float
+    add_column :feature_estimates, :generalBuffer, :float
   end
 end
